requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
numpy==1.26.2
rich==13.7.0
hydra-core==1.3.2
polars==0.20.1
wandb==0.16.1
scikit-learn==1.3.2
captum==0.7.0
python-dotenv==1.0.0
pydantic==2.5.3
torch==2.1.1
transformers==4.38.1
datasets==2.17.1
textattack==0.3.9
gdown==5.2.0
openai==1.99.9
